+++ app/libs/tone.h
@@ -5,7 +5,7 @@
 #include "utils/pins.h"
 
 void tone_init(uint8_t pin_num, TIMER32_TypeDef* _pwm_timer, TIMER32_TypeDef* _duration_timer);
-void tone(float freq, float duration);
+void tone(uint32_t freq, float duration);
 void set_position_change(uint8_t* _position);
 
 #endif
